fix: use correct format for favorite events
