+++ src/services/api/auth.service.ts
@@ -19,3 +19,27 @@ export const logout = async (): Promise<void> => {
 export const logoutAll = async (): Promise<void> => {
   await httpClient.post('/auth/logout-all')
 }
+
+/**
+ * Uploads an image to the server
+ * @param file The file to upload
+ * @param type The type of image: 'avatar' or 'banner' (default: 'avatar')
+ * @returns The public URL of the uploaded image
+ */
+export const uploadImage = async (
+  file: File,
+  type: 'avatar' | 'banner' = 'avatar'
+): Promise<string> => {
+  const formData = new FormData()
+  formData.append('file', file)
+  const response = await httpClient.post<{ url: string }>(
+    `/auth/upload?type=${type}`,
+    formData,
+    {
+      headers: {
+        'Content-Type': 'multipart/form-data'
+      }
+    }
+  )
+  return response.data.url
+}
